Extracted properties generation  to separate file in lib
